requirements.txt
requests
python-dotenv
pymongo
psycopg2-binary
black
pandas
sqlalchemy
isodate

